fix: use BRANCH when checking out viewer
Use the viewer that fits the nextcloud release.

Currently cypress tests are failing for all branches
even though the cause seems to be a recent commit to viewers stable branch.

Cypress runs for the stable23 branch of text
should use the stable23 branch of server
and the stable23 branch of viewer
so they match the actual releases that will go out.

Signed-off-by: Max <max@nextcloud.com>