use filelist PROPFIND for loading workspace
See #2171.

Even though this is meant to improve load times
I did not observe any improvements locally.

I suspect that is because the workspace was loaded in parallel with the PROPFIND anyway.
Experience may differ in production environments obviously.

It is still one request less and the code also seems more clean.

Signed-off-by: Max <max@nextcloud.com>